pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "wechat-plotkit"
version = "0.1.0"
description = "Codex-ready toolkit for extracting plotting styles from WeChat articles and reusing them in paper projects."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "beautifulsoup4>=4.12",
  "matplotlib>=3.8",
  "numpy>=1.26",
  "Pillow>=10.0",
  "playwright>=1.50",
  "requests>=2.32",
]

[project.optional-dependencies]
style = ["SciencePlots>=2.1"]
ocr = ["pytesseract>=0.3.10"]
dev = ["pytest>=8.2"]

[project.scripts]
wechat-plotkit = "wechat_plotkit.cli:main"

[project.urls]
Homepage = "https://github.com/handsomeZR-netizen/plotweaver-skillkit"
Repository = "https://github.com/handsomeZR-netizen/plotweaver-skillkit"
Issues = "https://github.com/handsomeZR-netizen/plotweaver-skillkit/issues"

[tool.setuptools.packages.find]
include = ["wechat_plotkit", "style_kit"]
